pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "singleclaw-dmn"
version = "0.2.0"
description = "SingleClaw DMN – a single-agent personal AI work OS with DMN memory, skill library, and guardian check."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
keywords = ["ai", "agent", "memory", "dmn", "cli", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "typer>=0.12",
    "pyyaml>=6.0",
    "rich>=13.0",
    "jsonschema>=4.20",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.scripts]
singleclaw = "singleclaw.cli:app"

[project.urls]
Homepage = "https://github.com/alicken-lai/SingleClaw-DMN"
"Bug Tracker" = "https://github.com/alicken-lai/SingleClaw-DMN/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["singleclaw*"]

[tool.ruff]
line-length = 99
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
